Return OptDefaults object to the correct heap
A recent change had allocated CmpContext::optDefaults_ from an NAHeap
but attempted to return it to the system heap. This resulted in
core-files when connection.close() is called and caused failures
in testing the T2 driver and the phoenix test. The change in this
commit properly returns the object to its NAHeap by correctly
using the NADELETE macro.

Change-Id: I1885313c2b7109f69f051efc53e04ef20d4c09d1
Closes-Bug: #1324220