fix for LaunchPad bug #1324299: compiler crashes in rangePartitioning
The issue is specific to the case when one of the key columns is of type
varchar. For this case, the internal logic that decodes the boundary
values expects the length to be at the beginning of the varchar. The
boundary values as read from hbase don't adhere to this condition.
Changes are done in NATable.cpp to update the boundary value to include
the length so the decoding logic can process it correctly.

Change-Id: Id77b0c71a410c03a6e54307302acb215e2a80657