Fix for Launch Pad bug 1274281
For select count(*) statement sql is not setting right query type
so result set access from t2 driver are failing.  The reason
sql is not setting right type is because when COPROC AGGR
operator replaced Scan operator the query type was not changed
and so query type was being set to the type of the COPROC AGGR
operator which is derived from UTILS operator.

Change-Id: Icb9f60e98a4182b780983d392430ceb692f68dd5