Classpath and install_local_hadoop changes.
Added Suresh's fix for Hive table access on Hortonworks systems.
Also made an equivalent fix for MapR systems, but was not able to
test it on a MapR system.

Improved (hopefully) detection of a Hortonworks system. We have seen
the HDP-changes file in /usr/lib/hadoop and in /usr/lib/hbase.

Fixed URLs in install_local_hadoop, so that the script can download
needed files from the Internet, without having the MY_LOCAL_SW_DIST
environment variable point to a local repository of tar files.
Note that it is still recommended to use $MY_LOCAL_SW_DIST.

Change-Id: Ic5528373edec98467d0614e3d4878b0dd99a4551