Fixing DTM early commit response to reply at correct point in state_change_commit_helper and sending the normal response at normal commit reponse at the end, in req_end_complete.
Additional changes to set the early commit response default to FALSE
(will NOT reply early, done for correctness).

Changes to set the default JVM allocation sizes, DTM set to 2048, SQL
set to 512.

Removing delay in sqlci that was added to have tests pass that were
failing due to an error 73.

Change-Id: I6baf0ad261d17d1ad9c227f31c423a4f26049a28