Fix for HBase 0.94.15 / Cloudera CDH 4.8.2
Updated the HBase-trx build to work for HBase 0.94.15, used in Cloudera
CDH 4.8.2, and possibly other versions. To do this, added a second
method to the HBase version used by Maven to build HBase-trx.
See https://wiki.trafodion.org/wiki/index.php/Building_the_Software#Notes
for more info on what modifications of HBase are needed.
Needed to change the version of this modified HBase to
hbase-0.94.6-trafodion-15 to get maven to pick up the changes.

Also fixed the HBase-trx Makefile to catch Maven errors. Those were
not detected, because Maven ran as first command in a pipe and we
need to set the pipefail option in the shell for that case.

Change-Id: I90faa5a7023ce0020f1e2fa00cc454871bd178f4