Fix for Bug #1325443
This fix addresses a recent regression which causes a select after an update to not see the newly updated values, when the select is part of the same transaction as the update. The problem was that executor methods were receiving both the old and new version, with the old version coming last, causing it to be version in the row used by the rest of the engine. The fix is to write a value into the executor row only if the timestamp value is greater than what may have already been processed for this column.

Change-Id: Ic976ca4b9219d284bfb55e75f4d4684968cfbdd7