Closes-Bug: 1328248 (INSERT ... SELECT query returned ERROR[8421] )
Error 8421 is reported when there is an attempt to store a NULL in
a NOT NULL column.  In this case, the user's query had properly
qualified the SELECT part of the query so that NULLs should not
have been passed to the INSERT part of the query ... and yet the
error was being reported.

This problem occurred only when the Native Expression feature was
enabled.  The fix is in the SQL Compiler code that generates native
code for an expression involving testing for NULLs.

The details of the fix and how/why it works are too complicated to put
into this commit message.  Reviewers who would like the details
should send email to the author of this fix and I will send an MS Word
document, complete with color coding, that will explain what I found
in the debugging process and how/why the fix solves the problem.

Change-Id: Ib5b89bbb4362487807c782c6614eef22b542aa43