Cardinality fix for queries against metadata tables
Cardinality estimation for some queries against metadata tables can be
improved. The specific case this fix is targeting is when the chosen
plan includes unique indices for access paths. In this case and even
though fake stats is currently being used for metadata tables,
cardinality of joins where the unique index is the right child is
better estimated to allow for more optimal (nested join) plans.

Change-Id: Id86dd41d9498979cb0fae98583cdf9d0b910e170
Closes-Gug: #1324303