Fixing early commit reply default to OFF.
There was a problem with the way the early commit flag was being
initialized.  So the default behavior would be unknown.  The change sets
the default of OFF to work and the DTM_EARLYCOMMITREPLY ms.env variable
can still be used to select desired behavior.

This change will fix many of the early commit issues we are seeing
including bug 1321498, 1305233, and executor regression tests.

Change-Id: I1a8c7bcae391cde522f9cda80db96aec80460eee