Multi-column stats based splitting does not work with the decimal data type
The multi-column stats based split algorithm used to assign data ranges to ESPs
does not work for the case where one of the salted table key columns is of
decimal data type. The reason was a previous restriction in the update stats logic
that prevented the generation of detailed stats for an MC (multi-column) stats if
any of the columns is decimal. That restriction is lifted with this fix.

When testing for this fix, another change related to the mc stats based
splitting was made to avoid asserting when the number of intervals is not
the same (is less) as the number of partitions in function
createRangePartitionBoundariesFromStats (in file NATable.cpp), but instead
return a null RangePartitionBoundaries. This issue was seen when the number of
intervals is smaller than the number of partitions.

Change-Id: I9b04c84bb1153a7d800794abdf8590029b7b828b
Closes-Bug: #1329977