Fix for bug 1330593, sqenvcom.sh when no mvn
Errors from "which mvn" now go to /dev/null.  This can occur on an
execution-only system where the Trafodion binaries were downloaded.

The bug was in part of sqenvcom.sh which is marked "BUILD Tools/Libraries".
That sets about 10 environment variables which are needed for building, but
not for execution-only systems. MAVEN is one of these, it has the
invocable name of the "mvn" executable.

A longer-term improvement would be to execute the approximately 30 lines
in the "BUILD Tools/Libraries" section of sqenvcom.sh only when building.
One test of this would be the presence of files which are only in the
full source, and not packaged as installable delivered files.

Change-Id: Ia21eaa126a4d0abbd787e9af4ac3b37fd6075c14