setup to access Hive tables on MapR clusters
There were a few things missing in the classpath for reading Hive
tables on MapR clusters. We have not yet been able to test this fully,
but these changes should get us closer. The two clusters we have seen
had their libhdfs.so files in different places. I could not find
a hadoop-mapreduce-client-core*.jar on the system, which is required
on other distributions. So far I've also not seen a related error
message to this missing jar, but we didn't get very far in our testing.

Change set 2: Addressed reviewer comments.

Change-Id: I4ffa3d7c5f1423059933127793604b51d89e570d