Fix for bug - Zookeeper not in connecting state
This is a fix for bug #1252790. The issue was in the timer that was
acting on the timed out CONNECTING state to move back to AVAILABLE was
actually processing an old connection. The fix now looks at the dialogue
IDs to ascertain that it is processing the current connection request.

Added checks for strtok output per Arvind's comment.

Change-Id: If0fad0da305258120157b9c9e3998be4225bd589