Fix for query cache issue for Hive selects.
Two changes
1)Fix for bug #1293816.
2)Discontinue linking in libprotobuf.so, since it is currently unused.

For the hive query cache bug, the issue was that any change to a HDFS file
in a Hive directory, or to the directory itself (add/drop a file), was not
reflected in the query cache key. So the compiler could give a plan with an
incorrect list of HDFS files to the HDFSScan operator. The fix is to add
max(fileInfo.mLastMod) to the query cache key. The max is taken over all
files for a given Hive table. The number of files for a given Hive table has
also been added to the query cache key to cover cases where a file is deleted
from a Hive directory. Both query cache and query text cache are addressed.
The mLastMod time for each file and the number of files are determined through
the libHdfs call hdfsListDirectory(), which we already make.

Linking in libprotobuf.so is causing issues on certain MapR clusters since
MapR also uses this library and sometimes the version used by MapR is
different from what Trafodion uses. Since this library is not being used by
Trafodion stack right now, we will no longer link in this library in SQL or
connectivity layers. When a fix is found for the version incompatibility
issue, this change will be reversed.

Patche Set 2.
Thank you Dave for catching these issues. They have been resolved in
Patch Set 2.

Change-Id: Idbe599a876fdcaf77d2bdb9fdbf4b77a3f431e46