Enable parallel ESP process creation
Several months ago, parallel ESP creation was disabled
and the checkin had the comment that this was being done
to fix a problem with some (unspecified) regression tests
failing. This change re-enables parallel ESP creation. I
have checked that no regression test failure occur because
of this change.

The change includes an unrelated testware fix for
core/TEST038. The problem was that writing to the test
log LOG038 from outside the sqlci session gives
unpredictable results, so this change removes these
kinds of writes to the test log.

Change-Id: I9cf1480e2ae2c52a35d9fafcc606b79743a74c20
Closes-Bug: #1336271