Fix bug 1336547 -arkesp aborting in Native Expr.
There was a place in the Native Expression code where we were picking up
either a 16-bit or a 32-bit value, assuming it was a positive number,
and using it as an offset into a buffer (in this case a string buffer).
The change is to use the 16-bit or 32-bit value as an unsigned offset
rather than signed.

Change-Id: I3eaac82c2580fe91179b7e50d4ada80bc1615cc0
modified:   sql/exp/ExpPCodeOptsNativeExpr.cpp