Retry more SQLCODEs after flushing metadata cache
This change adds more SQLCODEs to the list that are handled
by AQR, in an attempt to handle potential errors that can
happen when metadata is cached and DDL effecting the metadata
is performed concurrently. The new SQLCODEs are 1009, 2235,
4001,4023, and 4039. The AQR entries for these are configured
to flush metadata cache before the retry.

While adding these SQLCODEs to AQR, we noticed inconsistent
use of the AQRErrorMap::cmpInfo. Please see the email
attached to the associated launchpad bug report for an
explanation of the results of the investigation. As a result
of the investigation, this change also sets the value of
AQRErrorMap::cmpInfo to 0 for the new SQLCODEs mentioned
above, as well as for the other SQLCODEs that were using
the value 1, i.e., for SQLCODEs 1389 and 8581.

Closes-Bug: #1331083

Change-Id: Ib9d8cedfa7e0ac26f703258bd8347d028c6fae12