Removing sqf/conf/hbase-site.xml
This file doesn't have a complete set of HBase properties and using it
for HBase client or region server would mean that we would not pick
up other configuration items like zookeeper peer port or zookeeper
root, both of which are frequently set to non-default values.

After discussiong this with various people, including Marvin, we
decided to change the installer not to include $MY_SQROOT/conf in
the CLASSPATH and to get rid of this file. Both of these changes can
be made independently.

File sqf/sqenvcom.sh sets the class path to include the HBase conf
directory, as before.

Change-Id: Ie10da8f7cc64de0a609a7f35704379ab907bd43a