Fix for bug #1315537 - Dialogue ID does not match
The clients can sometimes get a Dialogue ID does not match error when
trying to connect. This was happening whenever both the
TerminateDialogue and BreakDialogue were being invoked, during the
previous client disconnect call, where the TerminateDialogue would make
the MXOSRVR state to AVAILABLE, invalidate the dialogue ID and
BreakDialogue would also attempt to do the same. The issue arises when
the latter call tries to change the state again but DCSMaster has
already allocated the MXOSRVR object reference to a new client.

Change-Id: I8e5a0bd5fd6a0827b4bd483d522f283c6d34348f