Fix for bug 1342141.
The change in plan for singleton select from serial to parallel is because of
recent checkin to bug 1333800. The root cause of this plan change is due to surround code.
This surround code doesn't make sense for trafodion where scan happens in
Hbase Regions server processes. Using count of CPUs from synthesized physical property
to scale-up already computed cost un-fairly penalizes serial plan.
It costs serial plan as if only one region server is doing the scan, if we do this,
serial plan will never win.

The Hbase scan cost is estimated for per probe, per partition (number of regions).
This is true for both serial and parallel plans.

The fix for this paln change involves removing surround code .
 The basic idea is to
 1. precisely model the work needed in HBase scan layer,
 2. normalize the cost per layer of ESPs, and
 3. sum the normalized cost over all ESP layers + HBase scan layer as the final cost.
 4. remove assumptions that are not true in Trafodion

Change-Id: Iccf98d8ec2f8ef677f3e0b209f2962a1939df130