Added files to be deleted during Cloudera uninstall
Added more files that need to be deleted during cloudera uninstall.

Added small "temporary" fix to trafodion_installer to check if user is
signed in as 'trafodion' to run the installer. This is only to prevent
issues in the short term while the installer team is working on a more
complete solution.

Edit: Added single quotes to prevent a wildcard issue.

Change-Id: I798dce1006c2ea5d8deceb4d7fceef6ef26d28a5