Made source directory name in 2 scripts match repo name
The directory name in SRCDIR now uses "win-odbc64" which
is the current name of the repository holding this source.
That reduces the amount of editing needed when building,
if the easiest "git clone ..." command was done.

The changed lines should have Windows EOLs (after a few tries), and the
license notice is changing to Windows EOLs in another branch.

Change-Id: I78bbcbc7f1296d24c7db35917cc52544e876aece