setting a cqd in sqlci does not have effect in embedded compiler
The changed set of CQDs in a cmp context is not propagated to another
cmp context when a switch takes place. The change ensures that the list
of CQDs that are changed in a cmp context (list kep in ControlDB) are
pushed to the cmp context we switch to. The change also ensures that the
list of reset CQDs are also kep in the ControlDB. The reason we need
this list now is that in case a CQD value was changed and was pushed its
value to a CI in a previous switch, we need to make sure that if the CQD
is reset, the default value is pushed during a switch.

Change-Id: I6519c4c99c086929f482d09759358b9ebf7f2171
Closes-Bug: #1347810