Fix for bug 1349542: error 4082 with T2 multi threaded programs
The problem is that when object ids are generated the code to get
the thread id is not executed for every thread because the code is
protected by a process global.  Because of this for every thread
other than first thread the first part of UID is 0 and that causes
problems which duplicate UIDs.  Fix is to replace the process
global with thread global.

Change-Id: Ia1d8a1811c505b805a89d628e20bf6129a4d14af