Adding support to use Hadoop 1.0 in a build environment
Steve Varnau found this issue when setting up a Hortonworks test machine.
HDP 1.3 uses Hadoop 1 interfaces and the libhdfs hdfsDelete() method
has only two parameters in Hadoop 1 while it has three parameters in
Hadoop 2.

Adding an environment variable to be able to use conditional compilation
for the calls to hdfsDelete().

In principle this means that we would need to produce two binaries, one
for Hadoop 1 and a second one for Hadoop 2, but I suspect that the same
binary will work ok for both. That's for two reasons: First, the value
of the extra parameter is irrelevant for our situation and, second, I
don't know when we actually delete an HDFS directory - maybe when we
do an INSERT OVERWRITE TABLE?

Change-Id: Ib5d17406a26c98abe39a2130a0cf852e8c4347bb