Fix to redube buffer sizes used by Hive scan.
A previous fix I had commited for this problem was causing certain bulk load
queries to not get a proper plan. Thanks to Khaled and Hans for uncovering
this issue. This change set undoes the previous change and implements the same
logic in the binder. This time the change is made only for the bulk loader.
If it is felt to be safe it can be extended for all inserts/upserts later.

Change-Id: I8e12a435008227aede6ba6bcd333db1dca2ba2f2