Added log files, bug fixes, changed setup scripts
Fixed return error bug.

Created log file directory on remote nodes so that log files can be
placed in them on the remote nodes.

Added log files to trafodion_setup and trafodion_installer.

Seperated hadoop distribution setup from trafodion setup.
traf_ambari_setup and traf_cloudera_setup are still avalible in the
/tools directory and will not run trafodion_setup.

EDIT 1: Removed options from trafodion_installer that are not needed.
Added check to make sure FQDN or IP address are not used as node names.
This is a check that will be needed until bug 1347971 is fixed, then
this check will need to be removed.

Change-Id: Iefb590e04434128e7deff3fdc0973be78ab41df2