Bug: #1354523
Added error display when 'hostname' string does not match the 'hostname'
string in the node section of the configuration file 'sqconfig'.
Added trace to display details of arguments passed to 'mpirun' when
creating monitor process to establish a Trafodion cluster.
Added checks to 'sqgen' scripts to error out when node section in
'sqconfig' is syntactically invalid.

Change-Id: I6ab1d85c06b168ad126dd32b72c83920cdfb6952