Reducing the path length for scan, single row get, batch get operations
Earlier, Key-values per row is buffered as a ByteBuffer and shipped to JNI side.There were 2 C->java transitions to get the Hbase row into tuple format.
Now, we avoid this copy on the  Java side and ship the length and offsets of
the Key-Value for all rows in a batch along with reference to the
Key-Value buffer. The data is directly moved from Hbase Buffers to
row format buffer. There are only 2 C->java transitions per batch of row.
This has shown 3-4 times reduction in path length in Trafodion SQL processes
 while scanning a table with 12 million rows.

Trafodion SQL processes dump core when ex_assert function is called or when it
exits with non-zero value.  These processes will also dump core when an
environment variable ABORT_ON_ERROR is set in release mode.

Fix for seabase/TEST010 failure with the patchSet 1

Rowset Select was dumping core sometimes. A wrong tuple was used to calculate the row length. It caused the proces to dump core sometimes.

Change-Id: I0ec4669b54971b6a0c699c0fa0662c85f68bd25d