Fix for bug 1348211 Generator error 7000 "root can't produce these values"
For joins involving repartitioning of data to match the other table,
we got this error in some cases, usually seen with salted tables.
Fix is to avoid storing the PIVs (partition input values) in the logical
scan node. Instead they are picked up in the preCodeGen phase, and this
means that if the PIVs have to be rewritten, we pick up the new and
correct values.

Patch set 3: Fixed regression issue by moving PIV logic so
availableValues includes the PIVs.

Change-Id: Ia8f9a83894e504f8d65d37d1589760258cdb8976