Fix for traf_cloudera_mods Hbase error, other fixes
Added check that will look to see what the exact name of hbase to use in
traf_cloudera_mods to make a call to set hbase settings and restart
hbase. (Formally clouderas url was .../hbase1 it is now .../hbase).

Changed parsing in trafodion_mods to allow for Cloudera cluster name to
be more than one word.

Moved SQconfig check in trafodion_installer to after it has been moved
to current directory if on a single node.

Changed error message in trafodion_mods.

Change-Id: I052b06bbe0eac8dadf70f5af051a16358aa35c15