Making Maven build output less verbose
Maven emits warnings about using an environment variable as the
version. We want to keep doing this, since we have multiple Maven
projects that should all have the same version, stored centrally
in the sqenvcom.sh file.

Suppressing most Maven output, including the warning, on stdout,
but preserving the entire output in the log file
(except for the "clean" target).

Change-Id: Ifee00e65971b20b485fc2eb65fff9135250d6fe3