Related to bug 1335477: Improve debugging code for MDAM.
If the Trafodion code is built using NA_MDAM_EXECUTOR_DEBUG, then
whenever a query uses an MDAM plan, the MDAM network will be
printed to cout when the query executes. This can be useful for
debugging MDAM queries.

Formerly, the code printed out pointers to the location of the
data values (and with the 64-bit conversion the pointers are
badly mangled). This is nearly useless for debugging.

With these changes, the code assumes the key column is a CHAR(10)
and prints out the actual value. (Note that there is no easy way
at present of determining the actual data type. If the data isn't
a CHAR(10), you'll get what you get; garbage most likely. But
since you're debugging anyway you can change the code to take
into account the actual data type of the example you are
debugging.)

Change-Id: I2d53159c1b10e28777916850efc3f325d3ddb234