Pre-fetch cells from Hbase
Pre-fetch is enabled via a parameter in HTableClient.startScan method.Pre-fetch
is not done for unique and batch Trafodion operations and all native
Hbase table access.  Pre-fetch is currently disabled for non-unique UMD
Trafodion operations.

startScan method invokes pre-fetch to Hbase in a different thread. When the
fetchRows method is called, pre-fetch completes, passes cell info to JNI and
invokes pre-fetch if there are more rows to be fetched.

We have observed around 45% reduction in response time to fetch  12 million
rows of a sizteen partition table in a node via a single process.

Change-Id: I3c81e182663fddd08a2fc873a39302b179850c92