Converting params of type ArrayList to array for Trafodion Java classses
Changed the parameters of the java functions like HTableClient.startGet,
startScan to use Array instead of ArrayList for the following reasons:
1.	Reduce the number of JNI to java transitions
2.	Reduce the path length in JNI
3.	Reduce the memory allocation from the system heap.

The first two items are expected to improve path length for select queries.
Single row select or small queries are expected to see this reduction as a
higher percentage because these methods are called only once per select query.
But, the percentage improvement is not measured.

3rd item in the above list is expected to reduce the memory usage in select
test when run for a longer time. It may not be due to leak in this area. But
when the memory allocation is intermixed between Trafodion code and the
faulty code, the block of memory as determined Memory Arena settings will be
left behind.  The trick is to reduce the memory allocation from the system heap in smaller units. The ArrayList objects were allocating memory from the
system heap.

Change-Id: I5a73456a6283a5fbf420b29f50895e5a007beab0