Externalized all of dcs.cloud.command to property file dcs-default.xml
Parts of the dcs.cloud.command were hard coded inside of DCS. With
this change all of the command is moved to the existing dcs.cloud.command
property contained in the DCS jar file resources. This allows the user
to manipulate the command without changing any DCS code. It's important
for the user manipulating the command to remember that the command must
return the result back in the form <internal ip addr>,<external id addr>
as DCS parses the string.

Change-Id: Icda9c688b63782a586382ae1f61f2f9c0b763174
Closes-Bug: #1370596