Fix for leak in Java Objects
The local references weren't explicitly  deleted in the JNI functions.
This caused the objects created in the JNI layer to leak. Fixed for all the
functions wherein StringArrayList and ByteArrayList were converted to use
array of objects. Need to investigate other functions as well.

Mike's comments are taken care of. THe comment from the previous submit is
also taken care of

Change-Id: I4c9d0c64c9bb8f4790411ebe80133bfb0944d6ed