Ensure UID returned for MD tables is non-null
When an NATable is created for a metadata table, the UID is not known.
Under certain conditions, this creates a problem whereby the AUTHS
table can not be read, and access to a table is incorrectly denied.
In this fix, the NATable::objectUid() function is changed to do a
lookup of the UID for a metadata table if it is 0, and to store the
result in NATable for future calls.

Change-Id: I5eb58871070e585a7242f92becde7e6a24a8f5f5
Closes-Bug: #1370289