Start RMS processes by default
This change configures instance startup to include an mxssmp and
mxsscp process to run on each node. The performance group has
verified that there is no performance regression when RMS processes
are running.

Even though RMS processes are running, the GET STATISTICS commands
are still not able to return statistics for a query id.

Change-Id: If6321b9e756b5ec94e274d5dc0194fe804973882
Implements: blueprint enable-rms-infrastructure