Fix for bug 1373591, 1370576, fix for port issue
In traf_ambari_mods the script is  now parsing for the port number and using
that to pass into the ambari configs.sh script to change HBase settings. The
script will otherwise use the default of 8080 which may be incorrect.

Bug 1373591 uninstall scripts are now removing distro repos.

Bug 1370576 cloudera and ambari setup scripts are now removing previous
installed files

Change-Id: I4160c8f0b9b246f13fb103465fe15566f541bfc2