Fix LP bug 1325716 -TIME(1) default CURRENT_TIME reported wrong values
When doing CSE (Common Subexpression Elimination), if we come across a
convert clause, we must interrogate the details of the conversion
before deciding this is a match with another conversion we have
already done.  In particular, the source's precision, scale, and type
must be the same as in the previous conversion and, likewise, for
the target.

Pre-reviewed by Justin.

All dev regressions were run to ensure that the fix has no
side-effects.

Files changed:
sql/exp/ExpPCodeOptimizations.cpp

Change-Id: I2705cef151ef163a43e1eef31ee47ef94d164051