Eliminate minor estimated nullcount inaccuracies
When the relative frequency of null values is estimated via
sampling when getting an estimated row count for an HBase
table, there is an (unlikely) situation in which the null
count could be thrown off. If the primary key consists of a
single column, and some row has all null values except for
the primary key, those nulls will be counted incorrectly.
This was caused by comparing two successive KeyValue positions
using < rather than <=.

In addition, if the end of the HFile is reached while taking
the sample, any nulls at the end of the last row will not be
counted, and this has been fixed as well.

o	Closes-Bug: #1383835

Change-Id: Ia449d1379d851e8df0f7811e835b5730851c33e2