LP Bug 1380733 - Phoenix tests fail with error 73
Analysis has shown that we are generating a commit request result of
COMMIT_CONFLICT in the transactional TrxRegionEndpoint coprocessor
for a transaction.  A COMMIT_CONFLICT result to a commit request
ultimately results in an error 73 being returned by SQL to the client.

In this case, a previous commit request, for the same transaction,
in the same region, had resulted in a successful commit analysis of COMMIT_OK.

The subsequent commit request, for the same transaction, is unexpected.
We are continuing to analyze the transactional prepare/commit process
to determine why we have this additional request.  For now, we have
added a workaround to the "hasConflict()" transactional processing
to recognize this condition and allow the initial conflict testing results
(in this case a COMMIT_OK) to be the returned result.

Change-Id: I8d35574384a0f64eac650a875827e19031bfd453