Make corrections to HFile path
To estimate the row count for an HBase table, the HFiles are
accessed directly from the HDFS file system. When the Trafodion
name of the table includes a delimited id, the quotes must not
be included in the node of the HDFS path representing the
qualified table name. In addition, the hbase.rootdir property
in hbase-site.xml may consist solely of a path rather than a
full URL. It was previously assumed that a full URL would be
present, and the value of the property was used to construct
a java.net.URI object. When a string consisting of only a file
path is passed to the URI constructor, a NullPointerException
is thrown (instead of the expected URISyntaxException), causing
the path we construct to the HFile to be incorrect. The code
was changed to utilize either a file path or a valid URI as the
value of the property.

Change-Id: If35d9da7aaab815a9c1d550bc505d86f0cbcf611
Closes-Bug: 1384959