Fixed external links and CSS stylesheet for the DCS docbook
Cherry picking changes submitted to proposed/0.8.3 by Susan

In the following XML source files in /dcs/src/docbkx, I updated
the external links (adding xlink:show="new") so that the links
now open in a separate window from the DCS book:

-- book.xml
-- configuration.xml
-- preface.xml
-- troubleshooting.xml

In book.xml, I removed unnecessary and confusing links from the
book's title and from the "Trafodion DCS" logo in the subtitle.
I changed the link in the Abstract so that the word "Trafodion"
links to the Trafodion wiki. I also added a missing period at
the end of that paragraph. I removed the question mark (?) from
the title "DCS" in the Overview section.

In preface.xml, I removed the link from "DCS version" to the
"Connectivity Subsystem" section in the Trafodion wiki (given
that the wiki content is subject to change, resulting in a broken
link) and replaced it with a more appropriate cross-reference to
the download site for the DCS package.

Per Arvind's advice, I updated the "Run modes" and "Example
Configurations" sections in configuration.xml to include the
alternate approach of specifying the host name and number of
servers.

I also updated the freebsd_docbook.css stylesheet in
/dcs/src/site/resources/css so that the DCS docbook automatically
appears in sans serif typeface (such as Arial), improving its
readability and look-and-feel.

(cherry picked from commit 4730742c501f5149fb453c9e7bfe5b4cc0e831c4)

Removed dos eol from freebsd_docbook.css as suggested by Susan.

Change-Id: Idd6f20123d628c7fd3a90b523af785575fa87ffe