Fix packaging of sql regress tests
The exclude file was excluding the sql/regress/tools/make-quoted-pattern.ksh
script used by UDR suite. This caused the test to fail when run outside of a
source tree.

Verified that removing this pattern does not change the number of files packaged
in the client and server product tar files.

Change-Id: I120bf32257a355abf06fad606a191d0fb11698ef