Use HBase estimated row count if no stats exist
In the absence of statistics, the row count for an HBase table
is estimated using information in its HFiles. Previouly, this
estimate was calculated and assigned to the NATable's
originalCardinality_ member variable, but this did not result
in it being used by the optimizer (the correct row count did
not appear in the results of an Explain).  The code to calculate
the estimate is now invoked from HSHbaseTableDef::getRowCount(),
which has the effect of it being used as intended.

Change-Id: I3794e8368082b6f17e47fe686ec9a004aed27976
Closes-Bug: #1388918