Uninstall command line parsing fix
Since a shift command occurs later in the loop, the --all option does
not need to to shift arguments. This caused a parsing error in the case
when the --all option is followed by other options.

Also removed quotes around a command line that caused shell to try to
find the entire command line as a single file name.

Change-Id: I0498c3aa7aa3dec4e4deca4d2466849feee77a6e