Fix for bug 1392090
Earlier JNI objects in Trafodion were made thread independent. HBaseClient_JNI
and HIveClient_JNI objects are  the creator/user of other Trafodion JNI objects.These objects couldn't be used in a different thread because the
thread specific JNIEnv was not setup. This change initializes the JNIENV when
it is not initialized in the thread.

Change-Id: I827002a495c3373b66fb2c88f80675d9da188193