Fix bug 1392522 - mxosrvr core dumped doing showplan (with N.E. enabled)
The basic problem was that PCodeCfg::generateShowPlan() called
PCodeCfg::layoutNativeCode() when the PCODE that had been generated
was unoptimized PCODE.  PCodeCfg::layoutNativeCode() expects to be
called with optimized PCODE and is not able to handle some strange
qualities about unoptimized PCODE.

This fix changes PCodeCfg::generateShowPlan() to put out a simple
message saying the a Native Expression exists but will not be
displayed as part of the showplan output.  In the future, we should
be able to implement a complete fix and display the Native Expression,
but in the meantime, we do not want to produce core dumps.

Since showplan is not documented for customer usage, this temporary
fix will not affect customers.  Also, if developers want to obtain
the assembler language for the Native Expression, we can obtain that
by turning on the CQDs that are used to debug the Native Expressions
logic.

Ran standard Trafodion developer regression tests.

Change-Id: I891a9c0bfbe680667a56f9b0ad4674ed0a0b1cbf